+++ docs/changelog.rst
@@ -35,7 +35,7 @@ Changed
 - Split ``WflowModel`` into ``WflowBaseModel`` containing all generic parts of all wflow models, like components and some setup functions, and ``WflowSbmModel``, containing the specific parts for the wflow_sbm concept.
 - ``WflowSedimentModel`` and ``WflowSbmModel`` both inherit from ``WflowBaseModel`` and extend it for their specific use cases.
 - The old ``WflowModel.setup_rivers`` has been split into three parts: ``WflowBaseModel.setup_rivers``, ``WflowSbmModel.setup_rivers`` and ``WflowSbmModel.setup_river_roughness``. Where the first contains the generic river setup logic, the second Sbm-specific implementations, and the third contains manning roughness implementations.
-
+- Clipping a model is now done by calling a single **clip** method. This function can be called with the ``hydromt update`` cli command.
 
 
 Unreleased
@@ -58,6 +58,7 @@ Fixed
 -----
 - **upgrade_to_v1_wflow**: fixed bug for [model] options that kept the same name in Wflow v1. (e.g. type, river_routing, land_routing). PR #487
 - **setup_floodplains**: states were not correctly added to the model config. PR #486
+- **setup_1d_model_connection**: improve clipping and snapping of 1D river with wflow basins/river. PR #416
 - Fix wflow build config example in the docs. PR #486
 - Fix crop_factor and water_frac values for grassland in CORINE. PR #523
 - included ``floodplain_water_flow__manning_n_parameter`` in naming script (#529)
@@ -103,7 +104,7 @@ Changed
 - **setup_lulcmaps** and equivalents: parameters to prepare from the mapping table are now linked to Wflow.jl variable names (dictionary and not list) to allow for renaming.  PR #364
 - **setup_output_config_timeseries**, **setup_outlets**, **setup_gauges**: the option to save parameters to netcdf scalar file as been renamed from `netcdf` to `netcdf_scalar` to better match the TOML file structure.  PR #364
 - Changed name of `g_tt` parameter to `g_ttm`, to align with the changes in https://github.com/Deltares/Wflow.jl/pull/512
-- **setup_soilmaps** [sediment]: add small and large aggregates to soil composition (additional to clay/silt/sand). Composition is now in fraction and not percentage. {R #331
+- **setup_soilmaps** [sediment]: add small and large aggregates to soil composition (additional to clay/silt/sand). Composition is now in fraction and not percentage. PR #331
 - **setup_soilmaps** [sediment]: additional parameters are prepared by the method (e.g. soil mean diameter, Govers transport capacity parameters). PR #331
 - **setup_constant_pars** [sediment]: added additional default values for sediment density and particle diameters. PR #331
 - **setup_riverbedsed** [sediment]: added option to derive Kodatie transport capacity parameters based on streamorder mapping. PR #331